\name{cn_keys}
\alias{cn_keys}
\docType{data}
\title{
Enem 2012 Natural Sciences Keys
}
\description{
The keys for the brazilian National High School Exam natural sciences test.
}
\usage{data("cn_keys")}
\format{
A data frame with 1 row and 45 cells, each cell containing the key to the corresponding. Values are numerical, ranging from 1 to 5.
}
\details{
The Brazilian National High School Exam (Exame Nacional do Esnsino Médio, Enem) is a facultative summative educational evaluation available for every Brazilian citizen that has finished, or is in the last year of high school. The test allows students to enroll in higher degree educational courses.

The exam is composed of 4 dichotomous tests: Humanities Sciences, Natural Sciences, Languages and Codes and Mathematics. Each test is composed of 45 multiple-choice items. In the year 2012, the test was taken by approximatelly 5.7 million examinees, whose data is freely available (see Source section).
}
\source{
\url{http://portal.inep.gov.br/basica-levantamentos-acessar}
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(cn_keys)
## maybe str(cn_keys) ; plot(cn_keys) ...
}
\keyword{datasets}
