% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize.responses.R
\name{normalize.responses}
\alias{normalize.responses}
\title{Normalize Range of Responses in a Matrix}
\usage{
normalize.responses(x, y)
}
\arguments{
\item{x}{A response matrix.}

\item{y}{A list containing the item keys.}
}
\value{
A \code{list} containing the normalized response matrix in the first
slot and its corresponding normalized keys in the second slot.
}
\description{
\code{normalize.responses} takes a response matrix and its keys
and puts its values between [1; n of keys for current item]
}
\details{
Although the use of a non-normalized response-matrix does not
interfere in the estimation of IRT parameters, sometimes it is desirable
to have item alternatives and its corresponding keys in a normalized range,
especially when performing distractor analysis with \code{\link{irt.report}}.

So, for example, a list composed of the values \code{[4, 3, 7]} would
become \code{[2, 1, 3]}.
}
\author{
Douglas De Rizzo Meneghetti
}
\keyword{normalization}
\keyword{responses}

