\name{dodoR-package}
\alias{dodoR-package}
\alias{dodoR}
\docType{package}
\title{
R functions I used for my master's thesis
}
\description{
This is a personal package I created in order to automate the generation of reports, charts and CSV files concerning the use of Item Response Theory in a binary matrix of responses.
}
\details{
\tabular{ll}{
Package: \tab dodoR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-30\cr
License: \tab What license is it under?\cr
}
}
\author{
Douglas De Rizzo Meneghetti

Maintainer: Douglas De Rizzo Meneghetti <douglasrizzo@fei.edu.br>
}
\references{

}
\keyword{ package }
\seealso{
\code{\link[mirt]{mirt}}
}
\examples{
%~~ simple examples of the most important functions ~~
}
