% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sample.rows.R
\name{sample.rows}
\alias{sample.rows}
\title{Random Samples of Rows}
\usage{
sample.rows(x, size, replace = F, prob = NULL)
}
\arguments{
\item{x}{Either a vector of one or more elements from which to choose, or a
positive integer.}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{Should sampling be with replacement?}

\item{prob}{A vector of probability weights for obtaining the elements of
the vector being sampled.}
}
\value{
A \code{data.frame} containing \code{size} rows extracted of
\code{x}.
}
\description{
\code{sample.rows} takes a sample of the specified size from the rows of x
using either with or without replacement
}
\details{
It basically does the same as \code{sample}, but returns row samples instead
of column samples.
}
\examples{
##Sample 5 random rows form the cars dataset
sample.rows(x = cars, size = 5)
}
\author{
Douglas De Rizzo Meneghetti
}
\references{
\url{http://stackoverflow.com/questions/8273313/random-rows-in-dataframe-in-r}
}
\keyword{rows}
\keyword{sample}

